week3/requirements.txt
fastapi
uvicorn
typer>=0.7.0
rich
click
chromadb
llama-index
llama-index-vector-stores-chroma
llama-index-embeddings-openai
llama-index-llms-openai
python-dotenv>=1.0.0